refactor flask app
### Description
Break up app.py into the flask app initialization and the routes.  In the long term, we will want to split up the routes into separate python modules and this simplifies that future step.

### Test plan
1. pytest
1. run locally
1. run on eb
