+++ Procfile
@@ -1,1 +1,1 @@
-web: PYTHONPATH=$(pwd)/src/py/:$(pwd)/third-party/sqlalchemy-enum-tables/ gunicorn --bind localhost:8000 --workers 1 --threads 15 aspen.app
+web: PYTHONPATH=$(pwd)/src/py/:$(pwd)/third-party/sqlalchemy-enum-tables/ gunicorn --bind localhost:8000 --workers 1 --threads 15 aspen.main
