+++ src/py/aspen/__main__.py
@@ -1,4 +1,4 @@
-from .app import application
+from .main import application
 
 # Setting debug to True enables debug output. This line should be
 # removed before deploying a production app.
