+++ src/py/aspen/main.py
@@ -1,0 +1,4 @@
+# this is where flask finds the entry point for the application.
+
+from aspen.app import views  # noqa: F401
+from aspen.app.app import application  # noqa: F401
